fix output tests + cleanup import cycles
