+++ docs/pluginreference/plugin_describe.md
@@ -1,0 +1,24 @@
+## plugin describe &mdash; Describe Plugin
+
+### Synopsis
+
+```
+plugin describe [<options>]
+```
+
+### Options
+
+```
+  -h, --help   help for describe
+```
+
+### Description
+
+Display a detailed description of the capabilities of this OCM plugin.
+
+### SEE ALSO
+
+##### Parents
+
+* [plugin](plugin.md)	 &mdash; OCM Plugin
+
