+++ pkg/contexts/ocm/plugin/cache/plugin.go
@@ -5,7 +5,7 @@
 package cache
 
 import (
-	"github.com/open-component-model/ocm/pkg/contexts/ocm/plugin/internal"
+	"github.com/open-component-model/ocm/pkg/contexts/ocm/plugin/descriptor"
 )
 
 type Plugin = *pluginImpl
@@ -14,14 +14,14 @@ type Plugin = *pluginImpl
 type pluginImpl struct {
 	name        string
 	source      *PluginSource
-	descriptor  *internal.Descriptor
+	descriptor  *descriptor.Descriptor
 	path        string
 	error       string
-	uploaders   *ConstraintRegistry[internal.UploaderDescriptor, internal.UploaderKey]
-	downloaders *ConstraintRegistry[internal.DownloaderDescriptor, internal.DownloaderKey]
+	uploaders   *ConstraintRegistry[descriptor.UploaderDescriptor, descriptor.UploaderKey]
+	downloaders *ConstraintRegistry[descriptor.DownloaderDescriptor, descriptor.DownloaderKey]
 }
 
-func NewPlugin(name string, path string, desc *internal.Descriptor, errmsg string) Plugin {
+func NewPlugin(name string, path string, desc *descriptor.Descriptor, errmsg string) Plugin {
 	p := &pluginImpl{
 		name:       name,
 		path:       path,
@@ -29,11 +29,11 @@ func NewPlugin(name string, path string, desc *internal.Descriptor, errmsg strin
 		error:      errmsg,
 	}
 	if desc != nil {
-		p.uploaders = NewConstraintRegistry[internal.UploaderDescriptor, internal.UploaderKey](desc.Uploaders)
-		p.downloaders = NewConstraintRegistry[internal.DownloaderDescriptor, internal.DownloaderKey](desc.Downloaders)
+		p.uploaders = NewConstraintRegistry[descriptor.UploaderDescriptor, descriptor.UploaderKey](desc.Uploaders)
+		p.downloaders = NewConstraintRegistry[descriptor.DownloaderDescriptor, descriptor.DownloaderKey](desc.Downloaders)
 	} else {
-		p.uploaders = NewConstraintRegistry[internal.UploaderDescriptor, internal.UploaderKey](nil)
-		p.downloaders = NewConstraintRegistry[internal.DownloaderDescriptor, internal.DownloaderKey](nil)
+		p.uploaders = NewConstraintRegistry[descriptor.UploaderDescriptor, descriptor.UploaderKey](nil)
+		p.downloaders = NewConstraintRegistry[descriptor.DownloaderDescriptor, descriptor.DownloaderKey](nil)
 	}
 	return p
 }
@@ -42,7 +42,7 @@ func (p *pluginImpl) GetSource() *PluginSource {
 	return p.source
 }
 
-func (p *pluginImpl) GetDescriptor() *internal.Descriptor {
+func (p *pluginImpl) GetDescriptor() *descriptor.Descriptor {
 	return p.descriptor
 }
 
@@ -69,7 +69,7 @@ func (p *pluginImpl) Error() string {
 	return p.error
 }
 
-func (p *pluginImpl) GetActionDescriptor(name string) *internal.ActionDescriptor {
+func (p *pluginImpl) GetActionDescriptor(name string) *descriptor.ActionDescriptor {
 	if !p.IsValid() {
 		return nil
 	}
@@ -82,12 +82,12 @@ func (p *pluginImpl) GetActionDescriptor(name string) *internal.ActionDescriptor
 	return nil
 }
 
-func (p *pluginImpl) GetAccessMethodDescriptor(name, version string) *internal.AccessMethodDescriptor {
+func (p *pluginImpl) GetAccessMethodDescriptor(name, version string) *descriptor.AccessMethodDescriptor {
 	if !p.IsValid() {
 		return nil
 	}
 
-	var fallback internal.AccessMethodDescriptor
+	var fallback descriptor.AccessMethodDescriptor
 	fallbackFound := false
 	for _, m := range p.descriptor.AccessMethods {
 		if m.Name == name {
@@ -106,50 +106,50 @@ func (p *pluginImpl) GetAccessMethodDescriptor(name, version string) *internal.A
 	return nil
 }
 
-func (p *pluginImpl) LookupDownloader(name string, artType, mediaType string) internal.List[*internal.DownloaderDescriptor] {
+func (p *pluginImpl) LookupDownloader(name string, artType, mediaType string) descriptor.List[*descriptor.DownloaderDescriptor] {
 	if !p.IsValid() {
 		return nil
 	}
 
-	return p.downloaders.LookupFor(name, internal.NewDownloaderKey(artType, mediaType))
+	return p.downloaders.LookupFor(name, descriptor.NewDownloaderKey(artType, mediaType))
 }
 
-func (p *pluginImpl) GetDownloaderDescriptor(name string) *internal.DownloaderDescriptor {
+func (p *pluginImpl) GetDownloaderDescriptor(name string) *descriptor.DownloaderDescriptor {
 	if !p.IsValid() {
 		return nil
 	}
 	return p.descriptor.Downloaders.Get(name)
 }
 
-func (p *pluginImpl) LookupUploader(name string, artType, mediaType string) internal.List[*internal.UploaderDescriptor] {
+func (p *pluginImpl) LookupUploader(name string, artType, mediaType string) descriptor.List[*descriptor.UploaderDescriptor] {
 	if !p.IsValid() {
 		return nil
 	}
 
-	return p.uploaders.LookupFor(name, internal.UploaderKey{}.SetArtifact(artType, mediaType))
+	return p.uploaders.LookupFor(name, descriptor.UploaderKey{}.SetArtifact(artType, mediaType))
 }
 
-func (p *pluginImpl) LookupUploadersForKeys(name string, keys internal.UploaderKeySet) internal.List[*internal.UploaderDescriptor] {
+func (p *pluginImpl) LookupUploadersForKeys(name string, keys descriptor.UploaderKeySet) descriptor.List[*descriptor.UploaderDescriptor] {
 	if !p.IsValid() {
 		return nil
 	}
 
-	var r internal.List[*internal.UploaderDescriptor]
+	var r descriptor.List[*descriptor.UploaderDescriptor]
 	for k := range keys {
 		r = r.MergeWith(p.uploaders.LookupFor(name, k))
 	}
 	return r
 }
 
-func (p *pluginImpl) LookupUploaderKeys(name string, artType, mediaType string) internal.UploaderKeySet {
+func (p *pluginImpl) LookupUploaderKeys(name string, artType, mediaType string) descriptor.UploaderKeySet {
 	if !p.IsValid() {
 		return nil
 	}
 
-	return p.uploaders.LookupKeysFor(name, internal.UploaderKey{}.SetArtifact(artType, mediaType))
+	return p.uploaders.LookupKeysFor(name, descriptor.UploaderKey{}.SetArtifact(artType, mediaType))
 }
 
-func (p *pluginImpl) GetUploaderDescriptor(name string) *internal.UploaderDescriptor {
+func (p *pluginImpl) GetUploaderDescriptor(name string) *descriptor.UploaderDescriptor {
 	if !p.IsValid() {
 		return nil
 	}
