+++ pkg/contexts/ocm/plugin/ppi/cmds/app.go
@@ -16,6 +16,7 @@ import (
 	"github.com/open-component-model/ocm/pkg/contexts/ocm/plugin/ppi"
 	"github.com/open-component-model/ocm/pkg/contexts/ocm/plugin/ppi/cmds/accessmethod"
 	"github.com/open-component-model/ocm/pkg/contexts/ocm/plugin/ppi/cmds/action"
+	"github.com/open-component-model/ocm/pkg/contexts/ocm/plugin/ppi/cmds/describe"
 	"github.com/open-component-model/ocm/pkg/contexts/ocm/plugin/ppi/cmds/download"
 	"github.com/open-component-model/ocm/pkg/contexts/ocm/plugin/ppi/cmds/info"
 	"github.com/open-component-model/ocm/pkg/contexts/ocm/plugin/ppi/cmds/topics/descriptor"
@@ -56,6 +57,7 @@ func NewPluginCommand(p ppi.Plugin) *PluginCommand {
 
 	cobrautils.TweakCommand(cmd, nil)
 
+	cmd.AddCommand(describe.New(p))
 	cmd.AddCommand(info.New(p))
 	cmd.AddCommand(action.New(p))
 	cmd.AddCommand(accessmethod.New(p))
